:hammer: split functions
